refactor(settings): show available integrations on the page instead of in a modal
The add-integration modal borrowed a pattern from directories with hundreds of
integrations. There are six, so hiding the unconnected ones cost a click and left
the page half empty below a short connected list.

Both groups are now always on the page: connected as rows with their identity and
health, available as a compact grid of tiles filling the space underneath. The
search field and capability grouping go with the modal, since neither earns its
structure across four tiles.

Connecting everything replaces the grid with a single line rather than an empty
region, and with nothing connected the page is just the available grid under a
"Get started" heading.
